+++ sample/src/main/java/digiteka/videofeed/andro/SampleApplication.kt
@@ -1,8 +1,8 @@
 package digiteka.videofeed.andro
 
 import android.app.Application
-import com.digiteka.newssnack.NewsSnack
-import com.digiteka.newssnack.core.config.DTKNSConfig
+import com.digiteka.videofeed.VideoFeed
+import com.digiteka.videofeed.core.config.DTKNSConfig
 import digiteka.videofeed.andro.logic.log.MyLogger
 import digiteka.videofeed.andro.logic.tracking.MyTracker
 
@@ -13,17 +13,17 @@ class SampleApplication : Application() {
 
 		//Create the DTKNSConfig
 		val dtknsConfig: DTKNSConfig = DTKNSConfig.Builder(
-			mdtk = BuildConfig.DIGITEKA_NEWSSNACK_MDTK
+			mdtk = BuildConfig.DIGITEKA_VIDEOFEED_MDTK
 		)
 			.build()
 
 		//Then initialize the NewsSnack sdk
-		NewsSnack.initialize(applicationContext = this, config = dtknsConfig)
+		VideoFeed.initialize(applicationContext = this, config = dtknsConfig)
 
 		// (Optional) set a custom logger
-		NewsSnack.setLogger(MyLogger)
+		VideoFeed.setLogger(MyLogger)
 
 		// (Optional) add a custom tracker
-		NewsSnack.addTracker(MyTracker)
+		VideoFeed.addTracker(MyTracker)
 	}
 }