+++ sample/src/main/java/digiteka/videofeed/andro/ui/samples/VingtMinutesActivity.kt
@@ -2,8 +2,8 @@ package digiteka.videofeed.andro.ui.samples
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.digiteka.newssnack.core.config.ui.DTKNSUiConfig
-import com.digiteka.newssnack.ui.NewsSnackFragment
+import com.digiteka.videofeed.core.config.ui.DTKNSUiConfig
+import com.digiteka.videofeed.ui.VideoFeedFragment
 import digiteka.videofeed.andro.R
 import digiteka.videofeed.andro.databinding.VingtMinutesActivityBinding
 
@@ -20,12 +20,24 @@ class VingtMinutesActivity : AppCompatActivity() {
 			.setPlayIcon(R.drawable.ic_vm_play)
 			.setPauseIcon(R.drawable.ic_vm_pause)
 			.setTitleFont(R.font.source_serif)
+			.setForcedDarkMode(true)
 			.setDescriptionFont(R.font.oswald)
 			.setZoneFont(R.font.arial)
+
+			.setDeactivateAds(false)
+			.setTagParams { zoneName, adId ->
+				mapOf(
+					"category" to "$zoneName _ $adId",
+					"sub_category" to "news"
+				)
+			}
+
 			.build()
 
-		binding.vingtMinutesFragmentContainerView.getFragment<NewsSnackFragment>()
-			.setUiConfig(uiConfig)
+		binding.vingtMinutesFragmentContainerView.getFragment<VideoFeedFragment>().apply {
+			setInjector(SampleInjector())
+			setUiConfig(uiConfig)
+		}
 
 	}
 }