+++ Changelog.md
@@ -1,4 +1,11 @@
 # Changelog
 
+## 1.1.0
+- updated videofeed sdk to version 1.1.0, and updated examples with new features: 
+  - `DTKNSViewInjector`;
+  - `DTKNSUiConfig.setDeactivateAds`;
+  - `DTKNSUiConfig.setTagParams`;
+  - And `DTKNSUiConfig.setZoneFont`.
+
 ## 1.0.3
 - Initial release of the sample application of VideoFeed